# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fonoteca.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/files/upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Write one file into the library at the given path.",
        "description": "The request body is the file itself — no multipart, so nothing is buffered to a temporary copy on the way in and an album-sized file costs one write. Upload one file per request; the catalogue learns about them at the next scan. 'folder' is a folder that already exists and is taken as it is; 'name' is the file's own path — one segment from a file picker, 'Album/CD1/01.flac' from a directory one — and is sanitised, because it is the only half the browser invented.",
        "operationId": "UploadLibraryFile",
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "$ref": "#/components/schemas/Stream"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileOperation"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/catalogue/releases/{id}/cover": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "The album's cover, as stored in the catalogue.",
        "description": "The first request for an album with no stored cover fetches the Cover Art Archive's front at 500px and keeps it, so the archive is asked once per album. Where the archive holds no front, Qobuz is asked for the same record — by barcode where there is one, by a matched title and artist otherwise — and refuses rather than guesses. An album neither has a picture of answers 404 and is remembered as such for a week, then asked about again, so a sleeve either gains later is still picked up. Served under an ETag with `no-cache`, so a changed cover shows on the next view and an unchanged one costs a 304.",
        "operationId": "GetReleaseCover",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Use an uploaded picture as the album's cover.",
        "description": "The request body is the image itself, its Content-Type one of JPEG, PNG, GIF, WebP, BMP or AVIF — never SVG, which is a document that runs script. At most 10 MB.",
        "operationId": "UploadReleaseCover",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/releases/{id}/cover/options": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Every image the Cover Art Archive holds for this album, to choose from.",
        "operationId": "GetReleaseCoverOptions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseCoverOptions"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/releases/{id}/cover/archive/{imageId}": {
      "post": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Use this Cover Art Archive image as the album's cover.",
        "operationId": "ChooseReleaseCover",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "imageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FileOperation": {
        "required": [
          "applied",
          "entries",
          "catalogueRows",
          "destination",
          "detail"
        ],
        "type": "object",
        "properties": {
          "applied": {
            "type": "boolean"
          },
          "entries": {
            "type": "integer",
            "format": "int32"
          },
          "catalogueRows": {
            "type": "integer",
            "format": "int32"
          },
          "destination": {
            "type": [
              "null",
              "string"
            ]
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReleaseCoverOption": {
        "required": [
          "id",
          "front",
          "types",
          "comment"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "front": {
            "type": "boolean"
          },
          "types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "comment": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ReleaseCoverOptions": {
        "required": [
          "mbid",
          "chosen",
          "uploaded",
          "qobuzAlbum",
          "images"
        ],
        "type": "object",
        "properties": {
          "mbid": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "chosen": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "uploaded": {
            "type": "boolean"
          },
          "qobuzAlbum": {
            "type": [
              "null",
              "string"
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseCoverOption"
            }
          }
        }
      },
...
      "Stream": {
        "type": "string",
        "format": "binary"
      },
...
    }
  },
...
}